+++ packages/playground/webpack.config.js
@@ -1,4 +1,4 @@
-const { makeConfig } = require('./webpack');
+const { makeConfig } = require('./webpack.makeConfig');
 
 module.exports = makeConfig({
   plainDevtool: Boolean(process.env.PLAIN_DEVTOOL),